LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/accessing-other-browsing-contexts/indexed-browsing-contexts-02-expected.txt


FAIL Accessing child browsing contexts 1 assert_equals: The top browsing context should have 3 child browsing contexts. expected 3 but got 1
PASS Accessing child browsing contexts 2
FAIL Accessing child browsing contexts 3 assert_equals: The top browsing context should have 3 child browsing contexts. expected 3 but got 1
